Option of deterministic keys and mnemonic for walletd
To create deterministic wallet new flag `--deterministic` is introduced. By default non-deterministic wallet is generated. In the case of deterministic wallet at first the secret spend key is generated for the first address, and secret spend key is derived from it. This first address will have a mnemonic seed which can be exported by the new RPC method `getMnemonicSeed` as a backup.

Added creation of the container from keys with flags `--view-key` and `--spend-key` which should be used along with `-g`.

Also container can be created from the mnemonic seed with `--mnemonic-seed` flag.