Fix of fee check in Core (correct rule and formatting) (#106)
This fixes syncing from zero with flag --without-checkpoints which means full check including PoW and fee that was borked in some of the attempts to make fee as compatible as possible with old wallets. Old fee v1is set to v2 in dirty hack to make sure that the use of currency.minimumFee() will result in enough fee for tx to be valid, so we will use MINIMUM_FEE_V1 definition from config.