Deterministic transaction keys (#120)
The `secret transaction key` allows, for example, proving that transaction was sent to a specified address and how much money this address received in a given transaction when used directly (undesired) or when used to generate a "proof of payment" (preferable). Since it is randomly generated, it needs to be stored in a wallet cache and/or recorded somewhere by the sender. In the case of wallet reset these keys are lost along with other transaction data, such as destination addresses. Likewise, upon restoring and importing the wallet from keys or mnemonic seed this transaction key is not recovered.

Replacing the randomly generated secret transaction key with deterministic using transaction `inputs` and sender's `secret view key` gives the next benefits:
* recovering the secret transaction key in case of wallet reset and/or restoring from mnemonic or keys
* recovering the secret transaction key even in tracking wallet thanks to using `secret view key` instead of `secret spend key`
* can be used to alert about spending in tracking wallet (possible further improvement of tracking wallets to enable tracking of outgoing transactions and displaying correct balance)